tam_dangnhap_doimk_quenmk_hd

+++ src/main/java/DomainModel/Combo.java
@@ -78,7 +78,9 @@ public class Combo implements Serializable {
 		return this.trangThai;
 	}
 
-	public void setTrangThai(byte trangThai) {
+	
+	public void setTrangThai(int trangThai) {
+
 		this.trangThai = trangThai;
 	}
